% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max-rectangle.R
\name{max_rect_under_histogram}
\alias{max_rect_under_histogram}
\title{Find the area of the maximum rectangle under the given histogram}
\usage{
max_rect_under_histogram(hts, ar_penalty = 0)
}
\arguments{
\item{hts}{histogram bar heights e.g. \code{c(2, 2, 1, 2, 2)}}

\item{ar_penalty}{Aspect ratio penalty (multiplication factor). Default value
of \code{0} means to return the largest area regardless of aspect ratio.   
Values above 0 are used to reduce the area by this factor multiplied
by the aspect ratio of the rectangle.
Value can be any 
positive floating point number. Useful range: 0-20}
}
\description{
Having the guts exposed in this function also help in manually tweaking the 
accepted "maximum area rectangle" by penalising areas we don't want.
Currently high-aspect-ratio areas are penalised so that areas which 
are more square might be chosen (even if actual total area is less)
}
\examples{
\dontrun{
hts <- c(2, 2, 1, 2, 2)
plot_hist_max_rect(hts, rect = NULL)
rect <- max_rect_under_histogram(hts)
plot_hist_max_rect(hts, rect = rect)

rect <- max_rect_under_histogram(hts, ar_penalty = 1)
plot_hist_max_rect(hts, rect = rect)
}

}
