% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition_matrix}
\alias{partition_matrix}
\title{Determine all rectangular regions of the same value in the matrix}
\usage{
partition_matrix(mat, ar_penalty = 0)
}
\arguments{
\item{mat}{matrix}

\item{ar_penalty}{Aspect ratio penalty (multiplication factor). Default value
of \code{0} means to return the largest area regardless of aspect ratio.   
Values above 0 are used to reduce the area by this factor multiplied
by the aspect ratio of the rectangle.
Value can be any 
positive floating point number. Useful range: 0-20}
}
\value{
data.frame
}
\description{
Determine all rectangular regions of the same value in the matrix
}
