% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition_matrix_single}
\alias{partition_matrix_single}
\title{Create a data.frame of rectangular areas of the target value in the matrix}
\usage{
partition_matrix_single(mat, target, ar_penalty = 0)
}
\arguments{
\item{mat}{matrix}

\item{target}{target value}

\item{ar_penalty}{Aspect ratio penalty (multiplication factor). Default value
of \code{0} means to return the largest area regardless of aspect ratio.   
Values above 0 are used to reduce the area by this factor multiplied
by the aspect ratio of the rectangle.
Value can be any 
positive floating point number. Useful range: 0-20}
}
\value{
data.frame
}
\description{
Create a data.frame of rectangular areas of the target value in the matrix
}
