% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max-rectangle.R
\name{plot_hist_max_rect}
\alias{plot_hist_max_rect}
\title{Plot a histogram and its max rectangle}
\usage{
plot_hist_max_rect(hts, rect = NULL)
}
\arguments{
\item{hts}{bar heights of histogram}

\item{rect}{output from \code{max_rect_under_histogram()}}
}
\value{
ggplot object
}
\description{
Plot a histogram and its max rectangle
}
