% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{plot_rects}
\alias{plot_rects}
\title{Plot all rects in a matrix}
\usage{
plot_rects(rects, mat = NULL)
}
\arguments{
\item{rects}{rects returned by \code{partition_matrix()}}

\item{mat}{matrix}
}
\value{
ggplot object
}
\description{
Plot all rects in a matrix
}
